+++ parent-project/scrapper/src/main/java/ru/tinkoff/edu/controller/LinksController.java
@@ -12,20 +12,20 @@ import java.util.List;
 @RestController()
 @RequestMapping("/links")
 public class LinksController {
-    private final LinkResponse defaultResponse = new LinkResponse(1, null);
+    private final LinkResponse defaultResponse = new LinkResponse(1L, null);
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ListLinksResponse getTrackedLinks(@RequestHeader("Tg-Chat-Id") Integer id) {
+    public ListLinksResponse getTrackedLinks(@RequestHeader("Tg-Chat-Id") Long id) {
         return new ListLinksResponse(List.of(defaultResponse), 1);
     }
 
     @PostMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public LinkResponse addTrackedLink(@RequestHeader("Tg-Chat-Id") Integer id, @RequestBody AddLinkRequest request) {
+    public LinkResponse addTrackedLink(@RequestHeader("Tg-Chat-Id") Long id, @RequestBody AddLinkRequest request) {
         return defaultResponse;
     }
 
     @DeleteMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public LinkResponse deleteTrackedLink(@RequestHeader("Tg-Chat-Id") Integer id, @RequestBody RemoveLinkRequest request) {
+    public LinkResponse deleteTrackedLink(@RequestHeader("Tg-Chat-Id") Long id, @RequestBody RemoveLinkRequest request) {
         return defaultResponse;
     }
 }
