+++ parent-project/scrapper/src/main/java/ru/tinkoff/edu/controller/TgChatController.java
@@ -7,12 +7,12 @@ import org.springframework.web.bind.annotation.*;
 public class TgChatController {
 
     @PostMapping(value = "/{id}")
-    public void registerChat(@PathVariable("id") Integer id) {
+    public void registerChat(@PathVariable("id") Long id) {
 
     }
 
     @DeleteMapping(value = "{id}")
-    public void deleteChat(@PathVariable("id") Integer id) {
+    public void deleteChat(@PathVariable("id") Long id) {
 
     }
 }
